refactor: PostEditDialog UI 분리
