+++ src/feature/post/ui/index.ts
@@ -3,3 +3,4 @@ export * from "./PostDetailButton"
 export * from "./PostEditButton"
 export * from "./PostDeleteButton"
 export * from "./PostAddDialog"
+export * from "./PostEditDialog"
